changes as per pr comments
